Fix FIND-PORT function
The FIND-PORT function may not find an available port from the given
range, but will return the last value of PORT variable, instead of NIL
which would indicate that no port in the range is available.

The change fixes that behaviour, so that NIL is returned when no ports
in the range is available.